hf-space-geolocalisation_retrieval_stockholm/requirements.txt
hopsworks==3.4.3
matplotlib==3.7.1
numpy==1.23.5
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
pandas==1.5.3
Pillow==10.2.0
protobuf==3.20.3
scikit_learn==1.2.2
tensorflow<=2.15.0
tensorflow_similarity==0.17.1
gitpython
plotly